requirements.txt
discord.py>=2.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0
aiohttp>=3.8.0
python-dotenv>=1.0.0